+++ django_dramatiq_charts/__init__.py
@@ -2,4 +2,4 @@
 # Project home page: https://github.com/ikvk/django_dramatiq_charts
 # License: Apache-2.0
 
-__version__ = '0.2.0'
+__version__ = '0.3.0'